+++ R/extraction.R
@@ -33,7 +33,8 @@ find_variable <- function(
     type = c("double", "integer", "character", "logical"),
     p_value = FALSE,
     variable = "",
-    exact = FALSE
+    exact = FALSE,
+    verbose = TRUE
 ) {
     if (!exact) {
         pattern <- gsub(
@@ -84,10 +85,16 @@ find_variable <- function(
 
     if (length(cols) == 0L) {
         # Aucune colonne
-        return(create_NA_type(type = type, len = nrow(demetra_m)))
+        return(list(
+            values = create_NA_type(type = type, len = nrow(demetra_m)),
+            missing = variable
+        ))
     } else if (length(cols) == 1L) {
         # 1 seule colonnes
-        return(demetra_m[, cols, drop = TRUE])
+        return(list(
+            values = demetra_m[, cols, drop = TRUE],
+            missing = NULL
+        ))
     }
 
     # Multiples colonnes
@@ -96,258 +103,174 @@ find_variable <- function(
     cond <- all(apply(
         X = demetra_m[, cols, drop = FALSE],
         MARGIN = 2L,
-        FUN = `==`,
+        FUN = identical,
         demetra_m[, cols[1L], drop = TRUE]
     ))
 
     if (cond) {
         id <- 1L
     } else {
-        message(
-            "Multiple column found for extraction of ",
-            variable,
-            "\n",
-            ifelse(p_value, "Last column selected", "First column selected")
-        )
+        if (verbose) {
+            message(
+                "Multiple column found for extraction of ",
+                toString(variable),
+                "\n",
+                ifelse(p_value, "Last column selected", "First column selected")
+            )
+        }
         id <- ifelse(p_value, length(cols), 1L)
     }
 
-    return(demetra_m[, cols[id], drop = TRUE])
+    return(list(
+        values = demetra_m[, cols[id], drop = TRUE],
+        missing = NULL
+    ))
 }
 
 extractStart <- function(demetra_m) {
-    missing_var <- NULL
-
     start_date <- find_variable(
         demetra_m,
         pattern = "(^span\\.start$)|(^start$)",
         type = "character",
-        variable = "start"
+        variable = "span.start"
     )
-    if (all(is.na(start_date))) {
-        missing_var <- "span.start"
-    }
-
-    return(list(values = start_date, missing = missing_var))
+    return(start_date)
 }
 
 extractEnd <- function(demetra_m) {
-    missing_var <- NULL
-
     end_date <- find_variable(
         demetra_m,
         pattern = "(^span\\.end$)|(^end$)",
         type = "character",
-        variable = "end"
+        variable = "span.end"
     )
-    if (all(is.na(end_date))) {
-        missing_var <- "span.end"
-    }
-
-    return(list(values = end_date, missing = missing_var))
+    return(end_date)
 }
 
 extractLog <- function(demetra_m) {
-    missing_var <- NULL
-
     log_transform <- find_variable(
         demetra_m,
         pattern = "^log$",
         type = "integer",
         variable = "log"
     )
-    if (all(is.na(log_transform))) {
-        missing_var <- c(missing_var, "log")
-    }
-
-    return(list(values = log_transform, missing = missing_var))
+    return(log_transform)
 }
 
 extractNobs <- function(demetra_m) {
-    missing_var <- NULL
-
     nobs <- find_variable(
         demetra_m,
         pattern = "(^span\\.n$)|(^n$)",
         type = "integer",
-        variable = "n"
+        variable = "span.n"
     )
-    if (all(is.na(nobs))) {
-        missing_var <- c(missing_var, "span.n")
-    }
-
-    return(list(values = nobs, missing = missing_var))
+    return(nobs)
 }
 
 extractNout <- function(demetra_m) {
-    missing_var <- NULL
-
     nout <- find_variable(
         demetra_m,
         pattern = "(^regression\\.nout$)|(^nout$)",
         type = "integer",
-        variable = "nout"
+        variable = "regression.nout"
     )
-    if (all(is.na(nout))) {
-        missing_var <- c(missing_var, "regression.nout")
-    }
-
-    return(list(values = nout, missing = missing_var))
+    return(nout)
 }
 
 extractNtd <- function(demetra_m) {
-    missing_var <- NULL
-
     ntd <- find_variable(
         demetra_m,
         pattern = "(^regression\\.ntd$)|(^ntd$)",
         type = "integer",
-        variable = "ntd"
+        variable = "regression.ntd"
     )
-    if (all(is.na(ntd))) {
-        missing_var <- c(missing_var, "regression.ntd")
-    }
-
-    return(list(values = ntd, missing = missing_var))
+    return(ntd)
 }
 
 extractM7 <- function(demetra_m) {
-    missing_var <- NULL
-
     m7 <- find_variable(
         demetra_m,
         pattern = "(^m\\.statistics\\.m7$)|(^m7$)",
         type = "double",
-        variable = "m7"
+        variable = "m-statistics.m7"
     )
-    if (all(is.na(m7))) {
-        missing_var <- c(missing_var, "m-statistics.m7")
-    }
-
-    return(list(values = m7, missing = missing_var))
+    return(m7)
 }
 
 extractMethod <- function(demetra_m) {
-    missing_var <- NULL
-
     m7 <- extractM7(demetra_m)
-    missing_var <- c(missing_var, m7$missing)
-
     method <- ifelse(is.na(m7$values), "Tramo-Seats", "X13")
-
-    return(list(values = method, missing = missing_var))
+    return(list(values = method, missing = m7$missing))
 }
 
 extractSeasonalFilter <- function(demetra_m) {
-    missing_var <- NULL
-
     seasonal_filters <- find_variable(
         demetra_m,
         pattern = "(^decomposition\\.seasonal\\.filters$)|(^seasonal\\.filters$)|(^decomposition\\.d9filter$)|(^decomposition\\.seasfilter$)|(^seasfilter$)",
         type = "character",
-        variable = "seasonal-filters"
+        variable = "decomposition.seasonal-filters"
     )
-    if (all(is.na(seasonal_filters))) {
-        missing_var <- c(missing_var, "decomposition.seasonal-filters")
-    }
-
-    return(list(values = seasonal_filters, missing = missing_var))
+    return(seasonal_filters)
 }
 
 extractTrendFilter <- function(demetra_m) {
-    missing_var <- NULL
-
     trend_filters <- find_variable(
         demetra_m,
         pattern = "(^decomposition\\.trend\\.filter$)|(^trend\\.filter$)|(^decomposition\\.d12filter$)|(^decomposition\\.trendfilter$)|(^trendfilter$)",
         type = "integer",
-        variable = "trend-filter"
+        variable = "decomposition.trend-filter"
     )
-    if (all(is.na(trend_filters))) {
-        missing_var <- c(missing_var, "decomposition.trend-filter")
-    }
-
-    return(list(values = trend_filters, missing = missing_var))
+    return(trend_filters)
 }
 
 extractQuality <- function(demetra_m) {
-    missing_var <- NULL
-
     quality <- find_variable(
         demetra_m,
         pattern = "(^quality\\.summary$)|(^summary$)|(^quality$)",
         type = "character",
-        variable = "quality"
+        variable = "quality.summary"
     )
-    if (all(is.na(quality))) {
-        missing_var <- c(missing_var, "quality.summary")
-    }
-
-    return(list(values = quality, missing = missing_var))
+    return(quality)
 }
 
 extractLeapYear <- function(demetra_m) {
-    missing_var <- NULL
-
     lp <- find_variable(
         demetra_m,
         pattern = "(^regression\\.lp$)|(^lp$)",
         type = "character",
-        variable = "lp"
+        variable = "regression.lp"
     )
-    if (all(is.na(lp))) {
-        missing_var <- c(missing_var, "regression.lp")
-    }
-
-    return(list(values = lp, missing = missing_var))
+    return(lp)
 }
 
 extractLeaster <- function(demetra_m) {
-    missing_var <- NULL
-
     leaster <- find_variable(
         demetra_m,
         pattern = "(^regression\\.leaster$)|(^leaster$)",
         type = "integer",
-        variable = "leaster"
+        variable = "regression.leaster"
     )
-    if (all(is.na(leaster))) {
-        missing_var <- c(missing_var, "regression.leaster")
-    }
-
-    return(list(values = leaster, missing = missing_var))
+    return(leaster)
 }
 
 extractAutoCorr <- function(demetra_m) {
-    missing_var <- NULL
-
     auto_corr <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.sa\\.ac1$)|(^seas\\.sa\\.ac1$)",
         type = "double",
-        variable = "seas-sa-ac1"
+        variable = "diagnostics.seas-sa-ac1"
     )
-    if (all(is.na(auto_corr))) {
-        missing_var <- c(missing_var, "diagnostics.seas-sa-ac1")
-    }
-
-    return(list(values = auto_corr, missing = missing_var))
+    return(auto_corr)
 }
 
 extractSeasCombined <- function(demetra_m) {
-    missing_var <- NULL
-
     presence_seasonality <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.lin\\.combined$)|(^seas\\.lin\\.combined$)",
         type = "character",
-        variable = "seas-sa-combined"
+        variable = "diagnostics.seas-lin-combined"
     )
-    if (all(is.na(presence_seasonality))) {
-        missing_var <- c(missing_var, "diagnostics.seas-lin-combined")
-    }
-
-    return(list(values = presence_seasonality, missing = missing_var))
+    return(presence_seasonality)
 }
 
 extract3Outliers <- function(demetra_m) {
@@ -385,41 +308,30 @@ extract3Outliers <- function(demetra_m) {
 }
 
 extractTDFTest <- function(demetra_m) {
-    missing_var <- NULL
-
     td_f_test <- find_variable(
         demetra_m,
         pattern = "(^regression\\.td\\.ftest$)|(^td\\.ftest$)",
         type = "double",
-        variable = "td-ftest",
+        variable = "regression.td-ftest:3",
         p_value = TRUE
     )
-    if (all(is.na(td_f_test))) {
-        missing_var <- c(missing_var, "regression.td-ftest:3")
-    }
-
-    return(list(values = td_f_test, missing = missing_var))
+    return(td_f_test)
 }
 
 extractIndependence <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
     independency_pvalue <- find_variable(
         demetra_m,
         pattern = "(^residuals\\.lb$)|(^lb$)|(^independence$)",
         type = "double",
-        variable = "independency",
+        variable = "residuals.lb:3",
         p_value = TRUE
     )
-    if (all(is.na(independency_pvalue))) {
-        missing_var <- c(missing_var, "residuals.lb:3")
-    }
 
     modalities <- cut(
-        x = as.numeric(independency_pvalue),
+        x = as.numeric(independency_pvalue$values),
         breaks = c(-Inf, thresholds[["residuals_independency"]]),
         labels = names(thresholds[["residuals_independency"]]),
         right = FALSE,
@@ -429,8 +341,8 @@ extractIndependence <- function(
 
     return(list(
         modalities = modalities,
-        values = independency_pvalue,
-        missing = missing_var
+        values = independency_pvalue$values,
+        missing = independency_pvalue$missing
     ))
 }
 
@@ -438,21 +350,16 @@ extractResidualsTDEffect <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
     td_ftest <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.td\\.sa\\.last$)|(^td\\.sa\\.last$)",
         type = "double",
-        variable = "td-ftest",
+        variable = "diagnostics.td-sa-last:2",
         p_value = TRUE
     )
-    if (all(is.na(td_ftest))) {
-        missing_var <- c(missing_var, "diagnostics.td-sa-last:2")
-    }
 
     modalities <- cut(
-        x = as.numeric(td_ftest),
+        x = as.numeric(td_ftest$values),
         breaks = c(-Inf, thresholds[["f_residual_td_on_sa"]]),
         labels = names(thresholds[["f_residual_td_on_sa"]]),
         right = FALSE,
@@ -462,27 +369,22 @@ extractResidualsTDEffect <- function(
 
     return(list(
         modalities = modalities,
-        values = td_ftest,
-        missing = missing_var
+        values = td_ftest$values,
+        missing = td_ftest$missing
     ))
 }
 
 extractResidualsSeasEffect <- function(demetra_m, thresholds = getOption("jdc_thresholds")) {
-    missing_var <- NULL
-
     residual_seasonality <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.sa\\.f$)|(^seas\\.sa.f$)",
         type = "double",
-        variable = "f_residual_s_on_sa",
+        variable = "diagnostics.seas-sa-f:2",
         p_value = TRUE
     )
-    if (all(is.na(residual_seasonality))) {
-        missing_var <- c(missing_var, "diagnostics.seas-sa-f:2")
-    }
 
     modalities <- cut(
-        x = as.numeric(residual_seasonality),
+        x = as.numeric(residual_seasonality$values),
         breaks = c(-Inf, thresholds[["f_residual_s_on_sa"]]),
         labels = names(thresholds[["f_residual_s_on_sa"]]),
         right = FALSE,
@@ -492,8 +394,8 @@ extractResidualsSeasEffect <- function(demetra_m, thresholds = getOption("jdc_th
 
     return(list(
         modalities = modalities,
-        values = residual_seasonality,
-        missing = missing_var
+        values = residual_seasonality$values,
+        missing = residual_seasonality$missing
     ))
 }
 
@@ -558,17 +460,12 @@ extractFrequency <- function(demetra_m) {
 }
 
 extractARIMA <- function(demetra_m) {
-    missing_var <- NULL
-
     arima_p <- find_variable(
         demetra_m,
         pattern = "(^arima\\.p$)|(^p$)",
         type = "integer",
         variable = "arima.p"
     )
-    if (all(is.na(arima_p))) {
-        missing_var <- c(missing_var, "arima.p")
-    }
 
     arima_d <- find_variable(
         demetra_m,
@@ -576,9 +473,6 @@ extractARIMA <- function(demetra_m) {
         type = "integer",
         variable = "arima.d"
     )
-    if (all(is.na(arima_d))) {
-        missing_var <- c(missing_var, "arima.d")
-    }
 
     arima_q <- find_variable(
         demetra_m,
@@ -586,9 +480,6 @@ extractARIMA <- function(demetra_m) {
         type = "integer",
         variable = "arima.q"
     )
-    if (all(is.na(arima_q))) {
-        missing_var <- c(missing_var, "arima.q")
-    }
 
     arima_bp <- find_variable(
         demetra_m,
@@ -596,9 +487,6 @@ extractARIMA <- function(demetra_m) {
         type = "integer",
         variable = "arima.bp"
     )
-    if (all(is.na(arima_bp))) {
-        missing_var <- c(missing_var, "arima.bp")
-    }
 
     arima_bd <- find_variable(
         demetra_m,
@@ -606,9 +494,6 @@ extractARIMA <- function(demetra_m) {
         type = "integer",
         variable = "arima.bd"
     )
-    if (all(is.na(arima_bd))) {
-        missing_var <- c(missing_var, "arima.bd")
-    }
 
     arima_bq <- find_variable(
         demetra_m,
@@ -616,17 +501,14 @@ extractARIMA <- function(demetra_m) {
         type = "integer",
         variable = "arima.bq"
     )
-    if (all(is.na(arima_bq))) {
-        missing_var <- c(missing_var, "arima.bq")
-    }
 
     arima_df <- data.frame(
-        arima_p = arima_p,
-        arima_d = arima_d,
-        arima_q = arima_q,
-        arima_bp = arima_bp,
-        arima_bd = arima_bd,
-        arima_bq = arima_bq
+        arima_p = arima_p$values,
+        arima_d = arima_d$values,
+        arima_q = arima_q$values,
+        arima_bp = arima_bp$values,
+        arima_bd = arima_bd$values,
+        arima_bq = arima_bq$values
     )
     arima_df[["arima_model"]] <- paste0(
         "(",
@@ -644,35 +526,37 @@ extractARIMA <- function(demetra_m) {
         arima_df[["arima_bq"]],
         ")"
     )
-    return(list(values = arima_df[["arima_model"]], missing = missing_var))
+    return(list(
+        values = arima_df[["arima_model"]],
+        missing = c(
+            arima_p$missing,
+            arima_d$missing,
+            arima_q$missing,
+            arima_bp$missing,
+            arima_bd$missing,
+            arima_bq$missing
+        )
+    ))
 }
 
 extractStatQ <- function(demetra_m, thresholds = getOption("jdc_thresholds")) {
-    missing_var <- NULL
-
-    q_value <- find_variable(
+    q_stat <- find_variable(
         demetra_m,
         pattern = "(^m\\.statistics\\.q$)|(^q$)",
         type = "double",
-        variable = "q statistic"
+        variable = "m-statistics.q"
     )
-    if (all(is.na(q_value))) {
-        missing_var <- c(missing_var, "m-statistics.q")
-    }
 
-    q_m2_value <- find_variable(
+    q_m2 <- find_variable(
         demetra_m,
         pattern = "(^m\\.statistics\\.q\\.m2$)|(^q\\.m2$)",
         type = "double",
-        variable = "q-m2 statistic"
+        variable = "m-statistics.q-m2"
     )
-    if (all(is.na(q_m2_value))) {
-        missing_var <- c(missing_var, "m-statistics.q-m2")
-    }
 
     stat_Q_modalities <- data.frame(
         q = cut(
-            x = as.double(q_value),
+            x = as.double(q_stat$values),
             breaks = c(-Inf, thresholds[["q"]]),
             labels = names(thresholds[["q"]]),
             right = FALSE,
@@ -680,7 +564,7 @@ extractStatQ <- function(demetra_m, thresholds = getOption("jdc_thresholds")) {
             ordered_result = TRUE
         ),
         q_m2 = cut(
-            x = as.double(q_m2_value),
+            x = as.double(q_m2$values),
             breaks = c(-Inf, thresholds[["q_m2"]]),
             labels = names(thresholds[["q_m2"]]),
             right = FALSE,
@@ -690,15 +574,18 @@ extractStatQ <- function(demetra_m, thresholds = getOption("jdc_thresholds")) {
         stringsAsFactors = FALSE
     )
     stat_Q_values <- data.frame(
-        q = as.double(q_value),
-        q_m2 = as.double(q_m2_value),
+        q = as.double(q_stat$values),
+        q_m2 = as.double(q_m2$values),
         stringsAsFactors = FALSE
     )
 
     return(list(
         modalities = stat_Q_modalities,
         values = stat_Q_values,
-        missing = missing_var
+        missing = c(
+            q_stat$missing,
+            q_m2$missing
+        )
     ))
 }
 
@@ -706,41 +593,31 @@ extractOOS_test <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
-    mean_value <- find_variable(
+    mean_stat <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.fcast\\.outsample\\.mean$)|(^mean$)",
         type = "double",
-        variable = "mean",
-        p_value = TRUE
-    )
-    if (all(is.na(mean_value))) {
-        missing_var <- c(
-            missing_var,
+        variable = c(
             "diagnostics.out-of-sample.mean:2",
             "diagnostics.fcast-outsample-mean:2"
-        )
-    }
+        ),
+        p_value = TRUE
+    )
 
-    mse_value <- find_variable(
+    mse_stat <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.fcast\\.outsample\\.variance$)|(^mse$)",
         type = "double",
-        variable = "mse",
-        p_value = TRUE
-    )
-    if (all(is.na(mse_value))) {
-        missing_var <- c(
-            missing_var,
+        variable = c(
             "diagnostics.out-of-sample.mse:2",
             "diagnostics.fcast-outsample-variance:2"
-        )
-    }
+        ),
+        p_value = TRUE
+    )
 
     stat_OOS_modalities <- data.frame(
         oos_mean = cut(
-            x = as.double(mean_value),
+            x = as.double(mean_stat$values),
             breaks = c(-Inf, thresholds[["oos_mean"]]),
             labels = names(thresholds[["oos_mean"]]),
             right = FALSE,
@@ -748,7 +625,7 @@ extractOOS_test <- function(
             ordered_result = TRUE
         ),
         oos_mse = cut(
-            x = as.double(mse_value),
+            x = as.double(mse_stat$values),
             breaks = c(-Inf, thresholds[["oos_mse"]]),
             labels = names(thresholds[["oos_mse"]]),
             right = FALSE,
@@ -759,15 +636,18 @@ extractOOS_test <- function(
     )
 
     stat_OOS_values <- data.frame(
-        oos_mean = as.double(mean_value),
-        oos_mse = as.double(mse_value),
+        oos_mean = as.double(mean_stat$values),
+        oos_mse = as.double(mse_stat$values),
         stringsAsFactors = FALSE
     )
 
     return(list(
         modalities = stat_OOS_modalities,
         values = stat_OOS_values,
-        missing = missing_var
+        missing = c(
+            mean_stat$missing,
+            mse_stat$missing
+        )
     ))
 }
 
@@ -775,62 +655,43 @@ extractDistributionTests <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
-    kurtosis_pvalue <- find_variable(
+    kurtosis_test <- find_variable(
         demetra_m,
         pattern = "(^residuals\\.kurtosis$)|(^kurtosis$)",
         type = "double",
-        variable = "kurtosis",
+        variable = "residuals.kurtosis:3",
         p_value = TRUE
     )
-    if (all(is.na(kurtosis_pvalue))) {
-        missing_var <- c(missing_var, "residuals.kurtosis:3")
-    }
 
-    skewness_pvalue <- find_variable(
+    skewness_test <- find_variable(
         demetra_m,
         pattern = "(^residuals\\.skewness$)|(^skewness$)",
         type = "double",
-        variable = "skewness",
+        variable = "residuals.skewness:3",
         p_value = TRUE
     )
-    if (all(is.na(skewness_pvalue))) {
-        missing_var <- c(missing_var, "residuals.skewness:3")
-    }
 
-    homoskedasticity_pvalue <- find_variable(
+    homoskedasticity_test <- find_variable(
         demetra_m,
         pattern = "(^residuals\\.lb2$)|(^lb2$)",
         type = "double",
-        variable = "homoskedasticity",
+        variable = "residuals.lb2:3",
         p_value = TRUE
     )
-    if (all(is.na(homoskedasticity_pvalue))) {
-        missing_var <- c(missing_var, "residuals.lb2:3")
-    }
 
-    independency_test_results <- extractIndependence(demetra_m, thresholds)
-    missing_var <- c(missing_var, independency_test_results$missing)
-    independency_pvalue <- independency_test_results$values
+    independency_test <- extractIndependence(demetra_m, thresholds)
 
-    normality_pvalue <- find_variable(
+    normality_test <- find_variable(
         demetra_m,
         pattern = "(^residuals\\.doornikhansen$)|(^doornikhansen$)|(^dh$)|(^normality$)",
         type = "double",
-        variable = "normality",
+        variable = c("residuals.dh:3", "residuals.doornikhansen:3"),
         p_value = TRUE
     )
-    if (all(is.na(normality_pvalue))) {
-        missing_var <- append(
-            x = missing_var,
-            values = c("residuals.dh:3", "residuals.doornikhansen:3")
-        )
-    }
 
     distribution_modalities <- data.frame(
         residuals_homoskedasticity = cut(
-            x = homoskedasticity_pvalue,
+            x = homoskedasticity_test$values,
             breaks = c(-Inf, thresholds[["residuals_normality"]]),
             labels = names(thresholds[["residuals_normality"]]),
             right = FALSE,
@@ -838,7 +699,7 @@ extractDistributionTests <- function(
             ordered_result = TRUE
         ),
         residuals_skewness = cut(
-            x = skewness_pvalue,
+            x = skewness_test$values,
             breaks = c(-Inf, thresholds[["residuals_skewness"]]),
             labels = names(thresholds[["residuals_skewness"]]),
             right = FALSE,
@@ -846,7 +707,7 @@ extractDistributionTests <- function(
             ordered_result = TRUE
         ),
         residuals_kurtosis = cut(
-            x = kurtosis_pvalue,
+            x = kurtosis_test$values,
             breaks = c(-Inf, thresholds[["residuals_kurtosis"]]),
             labels = names(thresholds[["residuals_kurtosis"]]),
             right = FALSE,
@@ -854,27 +715,33 @@ extractDistributionTests <- function(
             ordered_result = TRUE
         ),
         residuals_normality = cut(
-            x = normality_pvalue,
+            x = normality_test$values,
             breaks = c(-Inf, thresholds[["residuals_normality"]]),
             labels = names(thresholds[["residuals_normality"]]),
             right = FALSE,
             include.lowest = TRUE,
             ordered_result = TRUE
         ),
-        residuals_independency = independency_test_results$modalities
+        residuals_independency = independency_test$modalities
     )
     distribution_values <- data.frame(
-        residuals_homoskedasticity = homoskedasticity_pvalue,
-        residuals_skewness = skewness_pvalue,
-        residuals_kurtosis = kurtosis_pvalue,
-        residuals_normality = normality_pvalue,
-        residuals_independency = independency_pvalue
+        residuals_homoskedasticity = homoskedasticity_test$values,
+        residuals_skewness = skewness_test$values,
+        residuals_kurtosis = kurtosis_test$values,
+        residuals_normality = normality_test$values,
+        residuals_independency = independency_test$values
     )
 
     return(list(
         modalities = distribution_modalities,
         values = distribution_values,
-        missing = missing_var
+        missing = c(
+            homoskedasticity_test$missing,
+            skewness_test$missing,
+            kurtosis_test$missing,
+            normality_test$missing,
+            independency_test$missing
+        )
     ))
 }
 
@@ -882,25 +749,14 @@ extractOutliers <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
     nobs <- extractNobs(demetra_m)
-    missing_var <- c(missing_var, nobs$missing)
-    nobs <- nobs$values
-
     nout <- extractNout(demetra_m)
-    missing_var <- c(missing_var, nout$missing)
-    nout <- nout$values
-
     m7 <- extractM7(demetra_m)
-    missing_var <- c(missing_var, m7$missing)
-    m7 <- m7$values
-
-    pct_outliers_value <- 100.0 * nout / nobs
+    pct_outliers_value <- 100.0 * nout$values / nobs$values
 
     outliers_modalities <- data.frame(
         m7 = cut(
-            x = m7,
+            x = m7$values,
             breaks = c(-Inf, thresholds[["m7"]]),
             labels = names(thresholds[["m7"]]),
             right = FALSE,
@@ -917,14 +773,18 @@ extractOutliers <- function(
         )
     )
     outliers_values <- data.frame(
-        m7 = m7,
+        m7 = m7$values,
         pct_outliers = pct_outliers_value
     )
 
     return(list(
         modalities = outliers_modalities,
         values = outliers_values,
-        missing = missing_var
+        missing = c(
+            nobs$missing,
+            nout$missing,
+            m7$missing
+        )
     ))
 }
 
@@ -932,79 +792,49 @@ extractSeasTest <- function(
     demetra_m,
     thresholds = getOption("jdc_thresholds")
 ) {
-    missing_var <- NULL
-
     qs_residual_s_on_sa <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.sa\\.qs$)|(^seas\\.sa\\.qs$)",
         type = "double",
-        variable = "qs_residual_s_on_sa",
+        variable = c("diagnostics.seas-sa-qs:2", "diagnostics.seas-sa-qs"),
         p_value = TRUE
     )
-    if (all(is.na(qs_residual_s_on_sa))) {
-        missing_var <- append(
-            x = missing_var,
-            values = c("diagnostics.seas-sa-qs:2", "diagnostics.seas-sa-qs")
-        )
-    }
 
     qs_residual_sa_on_i <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.i\\.qs$)|(^seas\\.i\\.qs$)",
         type = "double",
-        variable = "qs_residual_sa_on_i",
+        variable = c("diagnostics.seas-i-qs:2", "diagnostics.seas-i-qs"),
         p_value = TRUE
     )
-    if (all(is.na(qs_residual_sa_on_i))) {
-        missing_var <- append(
-            x = missing_var,
-            values = c("diagnostics.seas-i-qs:2", "diagnostics.seas-i-qs")
-        )
-    }
 
     f_residual_s_on_sa <- extractResidualsSeasEffect(demetra_m)
-    missing_var <- c(missing_var, f_residual_s_on_sa$missing)
-    f_residual_s_on_sa <- f_residual_s_on_sa$values
 
     f_residual_sa_on_i <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.seas\\.i\\.f$)|(^seas\\.i\\.f$)",
         type = "double",
-        variable = "f_residual_sa_on_i",
+        variable = c("diagnostics.seas-i-f:2", "diagnostics.seas-i-f"),
         p_value = TRUE
     )
-    if (all(is.na(f_residual_sa_on_i))) {
-        missing_var <- append(
-            x = missing_var,
-            values = c("diagnostics.seas-i-f:2", "diagnostics.seas-i-f")
-        )
-    }
 
     f_residual_td_on_sa <- extractResidualsTDEffect(demetra_m)
-    missing_var <- c(missing_var, f_residual_td_on_sa$missing)
-    f_residual_td_on_sa <- f_residual_td_on_sa$values
 
     f_residual_td_on_i <- find_variable(
         demetra_m,
         pattern = "(^diagnostics\\.td\\.i\\.last$)|(^td\\.i\\.last$)",
         type = "double",
-        variable = "f_residual_td_on_i",
+        variable = c("diagnostics.td-i-last:2", "diagnostics.td-i-last"),
         p_value = TRUE
     )
-    if (all(is.na(f_residual_td_on_i))) {
-        missing_var <- append(
-            x = missing_var,
-            values = c("diagnostics.td-i-last:2", "diagnostics.td-i-last")
-        )
-    }
 
-    test_values <- cbind(
-        qs_residual_s_on_sa,
-        f_residual_s_on_sa,
-        qs_residual_sa_on_i,
-        f_residual_sa_on_i,
-        f_residual_td_on_sa,
-        f_residual_td_on_i
+    test_values <- data.frame(
+        qs_residual_s_on_sa = qs_residual_s_on_sa$values,
+        f_residual_s_on_sa = f_residual_s_on_sa$values,
+        qs_residual_sa_on_i = qs_residual_sa_on_i$values,
+        f_residual_sa_on_i = f_residual_sa_on_i$values,
+        f_residual_td_on_sa = f_residual_td_on_sa$values,
+        f_residual_td_on_i = f_residual_td_on_i$values
     )
 
     test_modalities <- as.data.frame(lapply(
@@ -1025,7 +855,14 @@ extractSeasTest <- function(
     return(list(
         modalities = test_modalities,
         values = test_values,
-        missing = missing_var
+        missing = c(
+            qs_residual_s_on_sa$missing,
+            f_residual_s_on_sa$missing,
+            qs_residual_sa_on_i$missing,
+            f_residual_sa_on_i$missing,
+            f_residual_td_on_sa$missing,
+            f_residual_td_on_i$missing
+        )
     ))
 }